Importer: Allow canceling an import
Previously the cancel action didn't persist. Now the cancel action will
clear out any local importer status objects and it will also send an API
request to stop the import on the server as well.

There were a few additional changes that had to be made to allow this to
work properly

 - `id` turned into `importerId` in the importer status object in the
   datastore
 - a new `CANCEL_PENDING` state exists for sending the stop signal to
   the server
 - `siteId` gets passed where it previously didn't
 - Update calling convention to the importer API endpoint