Importer: Add locking mechanism to freeze updates from the API
Previously, there have been cases where changes to the UI get
overwritten by changes coming from the API. For example, after
uploading an import file, if the user starts to map the authors, then a
successive update from the API (which contains no state transitions)
would reset the UI to the previous state.

This patch adds a locking system that the actions can use to freeze
things while the user makes UI updates. This locking system freezes
updates to the store until the lock is lifted. No updates are merged or
held in queue, rather they are simply dropped.

While dropping the updates could result in conflicts, it is unlikely to
be a real problem in actuality since this would require multiple people
trying to work on the same importer from different computers at the
same time.