Me: Add new section, "Get Apps"
This PR aims to add a new subsection under the me tag, subheadline "Special", labelled "Get Apps". The purpose is to create one central place to reference desktop and mobile application clients for WordPress and WordPress.com.

Mockup:

https://cloudup.com/c0BYBcoFwCu

Screenshot of current status:

https://cloudup.com/cLsipO-5bwx

Before I proceed, I would like feedback on this PR as-is. Both on content, design, and code, thank you. Depending on feedback, at some point I'll need to add a few images to this section, SVG or PNG in the case of app store link buttons. What would be the best approach for adding these images?